build(sbc): lint the motion engine and compile its other build paths
duet_motion was exempt from clang-tidy and forced -Wno-unused-parameter
on everything that included it, on the grounds that upstream lints its
own sources. The tree has diverged too far for that to be true, and the
exemption is what let the unreachable code accumulate.

The backlog was 26 findings, all fixed rather than suppressed. The two
that were judgement calls rather than mechanical: three methods that
never touched this become static, which is right here - unlike
MotionService::Configure, which was static only because it reached a
global - and class static members take an s_ prefix, which .clang-tidy
now states rather than leaving to fall through to the variable rule.

The build switches become #ifndef-guarded defaults so that -D can
override them, and CI compiles duet_motion with each one flipped. That is
the part that matters: a branch nobody compiles rots the first time a
field is renamed, which is exactly how these branches got into the state
this cleanup found them in. SUPPORT_S_CURVE and SUPPORT_LASER are left
out because they cannot build at all yet.

+++ src/DuetSbcInterface/src/Motion/DriveTracker.cpp
@@ -115,7 +115,7 @@ void DriveTracker::Advance(uint32_t now) noexcept
 			seg = m_segments; // EnterCurrentSegment may have merged it away
 		}
 
-		if ((uint32_t)(now - seg->GetStartTime()) < seg->GetDuration())
+		if ((now - seg->GetStartTime()) < seg->GetDuration())
 		{
 			return; // still running; GetCurrentPosition interpolates it
 		}
